Busy updating to using LineEditRV
